Fix up integration issues
