+++ src/Andy.Guard.Api/appsettings.Development.json
@@ -6,13 +6,12 @@
     }
   },
   "DownstreamApis": {
-    "PromptInjection": {
-      "BaseUrl": "https://promptscanner-dev.example/",
-      "RelativePath": "api/v1/scan",
+    "AndyInference": {
+      "BaseUrl": "http://localhost:5158/api",
       "Scopes": [
-        "api://promptscanner-dev/.default"
+        "api://cd.cip.andy-inference-api/.default"
       ],
-      "RequestAppToken": true
+      "RequestAppToken": false
     }
   }
 }
