+++ src/Andy.Guard.Api/appsettings.json
@@ -6,11 +6,11 @@
     }
   },
   "DownstreamApis": {
-    "PromptInjection": {
-      "BaseUrl": "https://promptscanner.example/",
-      "RelativePath": "api/v1/scan",
+    "AndyInference": {
+      "BaseUrl": "http://localhost:5158/api",
+      "RelativePath": "predict/batch",
       "Scopes": [
-        "api://promptscanner/.default"
+        "api://cd.cip.andy-inference-api/.default"
       ],
       "RequestAppToken": true
     }
