+++ tests/Andy.Guard.Tests/InferenceServiceFixture.cs
@@ -1,0 +1,211 @@
+using System.Net;
+using System.Reflection;
+using Andy.Guard.InputScanners;
+using Microsoft.Identity.Abstractions;
+using DotNet.Testcontainers.Builders;
+using DotNet.Testcontainers.Configurations;
+using DotNet.Testcontainers.Containers;
+using DotNet.Testcontainers.Networks;
+namespace Andy.Guard.Tests;
+
+public sealed class InferenceServiceFixture : IAsyncLifetime, IDisposable
+{
+    private const int TokenizerPort = 8000;
+    private const int InferencePort = 8080;
+
+    private readonly INetwork _network;
+    private readonly IContainer _tokenizerContainer;
+    private readonly IContainer _inferenceContainer;
+
+    private HttpClient? _httpClient;
+    private PromptInjectionScanner? _scanner;
+    private bool _disposed;
+
+    public InferenceServiceFixture()
+    {
+        _network = new NetworkBuilder()
+            .WithName($"andy-guard-tests-{Guid.NewGuid():N}")
+            .Build();
+
+        var configDirectory = ResolveConfigDirectory();
+        
+        _tokenizerContainer = new ContainerBuilder()
+            .WithImage("andy-inference-models-tokenizer-service:latest")
+            .WithName($"andy-guard-tokenizer-{Guid.NewGuid():N}")
+            .WithNetwork(_network)
+            .WithNetworkAliases("tokenizer-service")
+            .WithPortBinding(TokenizerPort, true)
+            .WithBindMount(configDirectory, "/app/config", AccessMode.ReadOnly)
+            .WithWaitStrategy(
+                Wait.ForUnixContainer()
+                    .UntilHttpRequestIsSucceeded(request => request
+                        .ForPort(TokenizerPort)
+                        .ForPath("/health")
+                        .ForStatusCode(HttpStatusCode.OK)))
+            .Build();
+
+        _inferenceContainer = new ContainerBuilder()
+            .WithImage("andy-inference-models-inference-service:latest")
+            .WithName($"andy-guard-inference-{Guid.NewGuid():N}")
+            .WithNetwork(_network)
+            .WithPortBinding(InferencePort, true)
+            .WithEnvironment("TokenizerServiceUrl", "http://tokenizer-service:8000")
+            .WithEnvironment("ModelConfiguration__TokenizerServiceUrl", "http://tokenizer-service:8000")
+            .WithWaitStrategy(
+                Wait.ForUnixContainer()
+                    .UntilHttpRequestIsSucceeded(request => request
+                        .ForPort(InferencePort)
+                        .ForPath("/health")
+                        .ForStatusCode(HttpStatusCode.OK)))
+            .Build();
+    }
+
+    public PromptInjectionScanner Scanner =>
+        _scanner ?? throw new InvalidOperationException("Fixture has not been initialized.");
+
+    public async Task InitializeAsync()
+    {
+        await _network.CreateAsync().ConfigureAwait(false);
+
+        await _tokenizerContainer.StartAsync().ConfigureAwait(false);
+        await _inferenceContainer.StartAsync().ConfigureAwait(false);
+
+        var hostPort = _inferenceContainer.GetMappedPublicPort(InferencePort);
+
+        _httpClient = new HttpClient
+        {
+            BaseAddress = new Uri($"http://localhost:{hostPort}/"),
+            Timeout = TimeSpan.FromSeconds(30)
+        };
+
+        var downstreamApi = CreateDownstreamApi(_httpClient);
+        var mapper = new PromptInjectionResultMapper();
+        _scanner = new PromptInjectionScanner(new InferenceApiClient(downstreamApi), mapper);
+    }
+
+    public async Task DisposeAsync()
+    {
+        await StopContainerAsync(_inferenceContainer).ConfigureAwait(false);
+        await StopContainerAsync(_tokenizerContainer).ConfigureAwait(false);
+
+        if (_network is not null)
+        {
+            await _network.DeleteAsync().ConfigureAwait(false);
+        }
+
+        _httpClient?.Dispose();
+    }
+
+    public void Dispose()
+    {
+        if (_disposed)
+            return;
+
+        _disposed = true;
+        _httpClient?.Dispose();
+    }
+
+    private static async Task StopContainerAsync(IContainer container)
+    {
+        try
+        {
+            await container.StopAsync().ConfigureAwait(false);
+        }
+        catch
+        {
+            // Ignore failures while stopping containers to keep test shutdown resilient.
+        }
+
+        await container.DisposeAsync().ConfigureAwait(false);
+    }
+
+    private static string ResolveConfigDirectory()
+    {
+        var configured = Environment.GetEnvironmentVariable("ANDY_INFERENCE_MODELS_CONFIG_PATH");
+        if (!string.IsNullOrWhiteSpace(configured) && Directory.Exists(configured))
+            return configured;
+
+        var defaultPath = Path.Combine(AppContext.BaseDirectory, "TestData", "inference-config");
+        if (Directory.Exists(defaultPath))
+            return defaultPath;
+
+        throw new DirectoryNotFoundException("Could not locate configuration directory for inference test containers.");
+    }
+
+    private static IDownstreamApi CreateDownstreamApi(HttpClient httpClient)
+    {
+        var proxy = DispatchProxy.Create<IDownstreamApi, DownstreamApiDispatchProxy>();
+        ((DownstreamApiDispatchProxy)(object)proxy!).Initialize(httpClient);
+        return proxy!;
+    }
+
+    private sealed class DownstreamApiDispatchProxy : DispatchProxy
+    {
+        private StaticDownstreamApi? _inner;
+
+        public void Initialize(HttpClient httpClient)
+        {
+            _inner = new StaticDownstreamApi(httpClient ?? throw new ArgumentNullException(nameof(httpClient)));
+        }
+
+        protected override object? Invoke(MethodInfo? targetMethod, object?[]? args)
+        {
+            if (_inner is null)
+                throw new InvalidOperationException("Downstream API proxy has not been initialized.");
+
+            if (targetMethod is not null &&
+                targetMethod.Name == nameof(IDownstreamApi.CallApiForAppAsync) &&
+                targetMethod.ReturnType == typeof(Task<HttpResponseMessage>))
+            {
+                var optionsDelegate = args is { Length: > 1 } ? args[1] as Delegate : null;
+                var content = args is { Length: > 2 } ? args[2] as HttpContent : null;
+                var cancellationToken = args is { Length: > 3 } && args[3] is CancellationToken token ? token : default;
+
+                return _inner.CallApiForAppAsync(optionsDelegate, content, cancellationToken);
+            }
+
+            throw new NotSupportedException($"Downstream API method '{targetMethod?.Name}' is not supported during integration tests.");
+        }
+    }
+
+    private sealed class StaticDownstreamApi
+    {
+        private readonly HttpClient _httpClient;
+
+        public StaticDownstreamApi(HttpClient httpClient)
+        {
+            _httpClient = httpClient;
+        }
+
+        public Task<HttpResponseMessage> CallApiForAppAsync(
+            Delegate? optionsDelegate,
+            HttpContent? content,
+            CancellationToken cancellationToken)
+        {
+            var options = new DownstreamApiOptions();
+            optionsDelegate?.DynamicInvoke(options);
+
+            var request = BuildRequest(options, content);
+            return _httpClient.SendAsync(request, cancellationToken);
+        }
+
+        private static HttpRequestMessage BuildRequest(DownstreamApiOptions options, HttpContent? content)
+        {
+            var method = options.HttpMethod switch
+            {
+                null or "" => HttpMethod.Get,
+                _ => new HttpMethod(options.HttpMethod)
+            };
+
+            var relativePath = options.RelativePath ?? string.Empty;
+            var requestUri = relativePath.Length > 0
+                ? new Uri(relativePath, UriKind.Relative)
+                : new Uri("/", UriKind.Relative);
+
+            var request = new HttpRequestMessage(method, requestUri) { Content = content };
+
+            options.CustomizeHttpRequestMessage?.Invoke(request);
+            return request;
+        }
+    }
+}
